+++ wsgid/commands/__init__.py
@@ -1,4 +1,5 @@
 
 
-import init
-import config
+from init import *
+from config import *
+from manage import *