+++ wsgid/loaders/__init__.py
@@ -19,7 +19,7 @@ class IAppLoader(plugnplay.Interface):
 
   '''
     Return the WSGI application object for tha app in app_path
-    @ap_path: Path on disk where the app is located. Already inserted into sys.path
+    @ap_path: Path on disk where the app is located. Already inserted into sys.path. This is --app-path + 'app/' 
     @app_full_name: Full qualified name for the WSGI application object
   '''
   def load_app(self, app_path, app_full_name):
@@ -38,6 +38,7 @@ def load_app(app_path, wsgi_app_full_name):
     return import_object(wsgi_app_full_name)
 
   app_loaders = IAppLoader.implementors()
+  log.debug("Found {0} loaders: {1}".format(len(app_loaders), app_loaders))
   for loader in app_loaders:
     if loader.can_load(app_path):
       log.info("Using AppLoader: %s" % loader.__class__.__name__)